edited snowbot
added a=variable in snowboydecoder in the start function, within the loop of the  start function a code is added to keep up the silence time since laast detection. if the max time is reached the added variable is called.
in the volume_+_demo_2 file cleaned up the code and added the extra functionality the new decoder offers